moved rewrite action tree view related methods from proof_panel to its own class
